Revert "Moves DoFnAdapters to runners-core"
This reverts commit 33ed3238e2b3899cff061be3056c5cc29fc60a04.
